+++ packages/xchain-crypto/rollup.config.js
@@ -1,6 +1,6 @@
 import commonjs from '@rollup/plugin-commonjs'
 import resolve from '@rollup/plugin-node-resolve'
-import typescript from 'rollup-plugin-typescript2'
+import typescript from '@rollup/plugin-typescript'
 import { visualizer } from 'rollup-plugin-visualizer'
 
 import pkg from './package.json'
@@ -32,6 +32,7 @@ export default {
   plugins: [
     resolve({ preferBuiltins: true, browser: true }),
     typescript({
+      declarationDir: 'lib',
       exclude: '__tests__/**',
     }),
     commonjs(),