+++ packages/xchain-dash/src/types/index.ts
@@ -1,1 +1,1 @@
-export { DashPreparedTx, NodeAuth, NodeUrls } from './client-types'
+export type { DashPreparedTx, NodeAuth, NodeUrls } from './client-types'